+++ x/evm/keeper/keeper.go
@@ -261,19 +261,13 @@ func (k *Keeper) GetAccountWithoutBalance(ctx sdk.Context, addr common.Address) 
 		return nil
 	}
 
-	codeHash := types.EmptyCodeHash
-	ethAcct, ok := acct.(evertypes.EthAccountI)
-	if ok {
-		codeHash = ethAcct.GetCodeHash().Bytes()
-	}
-
 	return &statedb.Account{
 		Nonce:    acct.GetSequence(),
-		CodeHash: codeHash,
+		CodeHash: k.GetCodeHash(ctx, addr.Bytes()).Bytes(),
 	}
 }
 
-// GetAccountOrEmpty returns empty account if not exist, returns error if it's not `EthAccount`
+// GetAccountOrEmpty returns empty account if not exist
 func (k *Keeper) GetAccountOrEmpty(ctx sdk.Context, addr common.Address) statedb.Account {
 	acct := k.GetAccount(ctx, addr)
 	if acct != nil {